I use astropy, astroquery and miriad-python to construct the sky model from an input miriad image. This then outputs a point source model, an ascii tab-separated version of the output table and a kvis annotation file which can be accessed in kvis by doing Overlay->Annotations. I've also used the cos^6 primary beam expression to attenuate the fluxes.
If you compare the points from the kvis annotation file and the model, you will see small offsets (a
few pixels). This is because of the projection effect. I am just using a simple cos(declination)
term to measure distances, which we need to work on. This may not be a problem for many fields where
the apparent flux is small, but may present a very real problem if the source ~100mJy.